+++ core/src/test/java/com/transferwise/common/gracefulshutdown/CustomSchedulerShutdownerIntTest.java
@@ -6,15 +6,15 @@ import static org.awaitility.Awaitility.await;
 import com.transferwise.common.gracefulshutdown.strategies.TaskSchedulersGracefulShutdownStrategy;
 import com.transferwise.common.gracefulshutdown.test.BaseTestEnvironment;
 import com.transferwise.common.gracefulshutdown.test.TestApplication;
-import com.transferwise.common.gracefulshutdown.test.TestBApplication;
+import com.transferwise.common.gracefulshutdown.testcustomscheduler.CustomSchedulerConfiguration;
 import lombok.SneakyThrows;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.test.context.ContextConfiguration;
 
 @BaseTestEnvironment
-@SpringBootTest(classes = {TestBApplication.class})
-class AlternativeSchedulingShutdownerIntTest {
+@ContextConfiguration(classes = {CustomSchedulerConfiguration.class})
+class CustomSchedulerShutdownerIntTest {
 
   @Autowired
   private GracefulShutdowner gracefulShutdowner;